% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_normal_data.R
\name{gen_normal_data}
\alias{gen_normal_data}
\title{gen_normal_data}
\usage{
gen_normal_data(n.seas, n.reg, weeks, mu, new.mu = NULL, sd1, sd2, tau = 1)
}
\arguments{
\item{n.seas}{number of seasons to simulate}

\item{n.reg}{number of regions to simulate}

\item{weeks}{number of weeks in a season}

\item{mu}{vector of means to be sampled from (w/ replacement)}

\item{new.mu}{specific mean to be used in the last season}

\item{sd1}{standard dev of data model}

\item{sd2}{standard dev of the normal functional form}

\item{tau}{scaling factor (default=1)}
}
\value{
A data.frame
}
\description{
A function to generate simulated seasons
}
\examples{
tmpd=gen_normal_data(n.seas=2, n.reg=2, weeks=1:30, mu=c(11,17), sd1=0.015, sd2=3)
library(ggplot2)
ggplot(data=tmpd) + geom_point(aes(x=week, y=y)) + 
facet_grid(region~season) + geom_line(aes(x=week,y=meanfc))

}
