% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmc_est.R
\name{get_mcmc_est}
\alias{get_mcmc_est}
\title{get_mcmc_est}
\usage{
get_mcmc_est(mcmc.fit, funct = "mean")
}
\arguments{
\item{mcmc.fit}{mcmc fit to grab the seasons}

\item{funct}{Either mean or median (default="mean")}
}
\value{
data.frame
}
\description{
grabs the fits from all sesons
}
\examples{


}
