% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmc_fit.R
\name{get_mcmc_fit}
\alias{get_mcmc_fit}
\title{get_mcmc_fit}
\usage{
get_mcmc_fit(mcmc.est, data)
}
\arguments{
\item{mcmc.est}{estimated values given by get_mcmc_est}

\item{data}{data.frame with season, region, weeks(time) variable required}
}
\value{
data.frame
}
\description{
A function which returns the fitted line based on mcmc est
}
\examples{
get_mle_fit(inits, data)

}
