% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mle_fit.R
\name{get_mle_fit}
\alias{get_mle_fit}
\title{get_mle_fit}
\usage{
get_mle_fit(inits, data)
}
\arguments{
\item{inits}{initial values given by}

\item{data}{data.frame with season, region, weeks(time) variable required}
}
\value{
data.frame
}
\description{
A function which returns the fitted line based on mle params
}
\examples{
get_mle_fit(inits, data)

}
