% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priors.R
\name{get_priors}
\alias{get_priors}
\title{get_priors}
\usage{
get_priors(mcmc.obj)
}
\arguments{
\item{mcmc.obj}{mcmc object}
}
\value{
list
}
\description{
Use mcmc objects from \code{\link[functform]{drake_mcmc}} to get priors for importance sampling
}
\examples{




}
