% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isForecast.R
\name{isForecast}
\alias{isForecast}
\title{isForecast}
\usage{
isForecast(y, n.weeks, seas.fits, priors, prior.type = "Unif", n.iter, tau,
  sigma, samp_type = "Transformation")
}
\arguments{
\item{y}{data from forecast season}

\item{n.weeks}{number of points from the forecast season to include}

\item{seas.fits}{matrix season fits to be used in sampling (see \code{\link{get_season_fit}})}

\item{priors}{priors from \code{\link{get_priors}} or vector of means}

\item{prior.type}{type of prior to use (default is "Unif)}

\item{n.iter}{number of iterations to run}

\item{tau}{scaling factor for est}

\item{sigma}{sd used in generating weights}

\item{samp_type}{Type of sampling procedure to be preformed}
}
\value{
list
}
\description{
Preforms both types of importance sampling forecasting
}
\examples{


}
\seealso{
\url{http://www.r-project.org}
}
